cleanup core/database/records/Game
