+++ atlas/core/database/Database.cpp
@@ -11,7 +11,7 @@
 #include "RapidTransaction.hpp"
 #include "core/config.hpp"
 #include "core/database/migrations/templates.hpp"
-#include "core/database/record/Game.hpp"
+#include "core/database/record/game/Game.hpp"
 
 namespace internal
 {
