+++ atlas/core/import/Importer.hpp
@@ -12,7 +12,7 @@
 #include <filesystem>
 
 #include "core/Types.hpp"
-#include "core/database/record/Game.hpp"
+#include "core/database/record/game/Game.hpp"
 
 struct GameImportData;
