+++ atlas/ui/delegates/RecordBannerDelegate.cpp
@@ -11,9 +11,9 @@
 #include <tracy/Tracy.hpp>
 
 #include "core/config.hpp"
-#include "core/database/record/Game.hpp"
 #include "core/database/record/GameData.hpp"
 #include "core/database/record/Version.hpp"
+#include "core/database/record/game/Game.hpp"
 #include "core/utils/QImageBlur.hpp"
 #include "ui/models/RecordListModel.hpp"
