+++ atlas/ui/models/RecordListModel.hpp
@@ -10,7 +10,7 @@
 
 #include <iostream>
 
-#include "core/database/record/Game.hpp"
+#include "core/database/record/game/Game.hpp"
 
 struct ImageLoader final : public QObject
 {
