+++ atlas/ui/views/RecordListView.hpp
@@ -8,8 +8,8 @@
 #include <QListView>
 
 #include "core/Types.hpp"
-#include "core/database/record/Game.hpp"
 #include "core/database/record/GameData.hpp"
+#include "core/database/record/game/Game.hpp"
 
 enum DelegateType
 {
