+++ atlas/ui/views/VersionView.hpp
@@ -7,9 +7,9 @@
 
 #include <QWidget>
 
-#include "core/database/record/Game.hpp"
 #include "core/database/record/GameData.hpp"
 #include "core/database/record/Version.hpp"
+#include "core/database/record/game/Game.hpp"
 
 QT_BEGIN_NAMESPACE
