+++ atlas/ui/widgets/games/GameWidget.hpp
@@ -3,8 +3,8 @@
 
 #include <QWidget>
 
-#include "core/database/record/Game.hpp"
 #include "core/database/record/GameData.hpp"
+#include "core/database/record/game/Game.hpp"
 
 QT_BEGIN_NAMESPACE
